+++ content.go
@@ -1,10 +1,13 @@
 package main
 
 import (
+	"encoding/json"
+	"errors"
 	"fmt"
 	"html/template"
 	"net/http"
 	"runtime"
+	"strings"
 
 	"github.com/prometheus/common/version"
 	"gopkg.in/yaml.v3"
@@ -44,6 +47,7 @@ const (
           <div><a href="{{ .MetricsPath }}">Metrics</a></div>
           <div><a href="/config">Configuration</a></div>
           <div><a href="/targets">Targets</a></div>
+          <div><a href="/loglevel">loglevel</a></div>
           <div><a href="/status">Status</a></div>
           <div><a href="/sql_exporter_metrics">Exporter Metrics</a></div>
           <div><a href="/debug/pprof">Profiling</a></div>
@@ -57,6 +61,10 @@ const (
     {{ define "content.home" -}}
       <p>This is a <a href="{{ .DocsUrl }}">Prometheus {{ .ExporterName }}</a> instance.
         You are probably looking for its <a href="{{ .MetricsPath }}">metrics</a> handler.</p>
+    {{- end }}
+
+	{{ define "content.health" -}}
+      <H2>{{ .Message }}</H2>
     {{- end }}
 
     {{ define "content.config" -}}
@@ -138,12 +146,17 @@ type tdata struct {
 
 	// status
 	Version versionInfo
+
+	// health - loglevel
+	Message string
+
 	// `/error` only
 	Err error
 }
 
 var (
 	allTemplates    = template.Must(template.New("").Parse(templates))
+	healthTemplate  = pageTemplate("health")
 	homeTemplate    = pageTemplate("home")
 	configTemplate  = pageTemplate("config")
 	targetsTemplate = pageTemplate("targets")
@@ -167,20 +180,145 @@ func HomeHandlerFunc(metricsPath string, exporter Exporter) func(http.ResponseWr
 	}
 }
 
+func HealthHandlerfunc(metricsPath string, exporter Exporter) func(http.ResponseWriter, *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var status []byte
+		accept_type := r.Header.Get(acceptHeader)
+		if strings.Contains(accept_type, applicationJSON) {
+			w.Header().Set(contentTypeHeader, applicationJSON)
+			status = []byte("{\"status\"=\"ok\"}")
+			w.Header().Set(contentLengthHeader, fmt.Sprint(len(status)))
+		} else if strings.Contains(accept_type, textPLAIN) {
+			w.Header().Set(contentTypeHeader, textPLAIN)
+			status = []byte("OK")
+		} else {
+			w.Header().Set(contentTypeHeader, textHTML)
+			healthTemplate.Execute(w, &tdata{
+				ExporterName: exporter.Config().Globals.ExporterName,
+				MetricsPath:  metricsPath,
+				DocsUrl:      docsUrl,
+				Message:      "OK",
+			})
+			return
+		}
+		w.Write(status)
+	}
+}
+
 // ConfigHandlerFunc is the HTTP handler for the `/config` page. It outputs the configuration marshaled in YAML format.
 func ConfigHandlerFunc(metricsPath string, exporter Exporter) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
-		config, err := exporter.Config().YAML()
-		if err != nil {
-			HandleError(0, err, metricsPath, exporter, w, r)
+		accept_type := r.Header.Get(acceptHeader)
+		if strings.Contains(accept_type, applicationJSON) {
+			conf, err := exporter.Config().JSON()
+			if err != nil {
+				HandleError(0, err, metricsPath, exporter, w, r)
+				return
+			}
+			w.Header().Set(contentTypeHeader, applicationJSON)
+			w.Header().Set(contentLengthHeader, fmt.Sprint(len(conf)))
+			w.Write(conf)
+		} else {
+			config, err := exporter.Config().YAML()
+			if err != nil {
+				HandleError(0, err, metricsPath, exporter, w, r)
+				return
+			}
+			configTemplate.Execute(w, &tdata{
+				ExporterName: exporter.Config().Globals.ExporterName,
+				MetricsPath:  metricsPath,
+				DocsUrl:      docsUrl,
+				Config:       string(config),
+			})
+		}
+	}
+}
+
+// LogLevelHandlerFunc is the HTTP handler for the POST loglevel entry point (`/loglevel`).
+func LogLevelHandlerFunc(metricsPath string, exporter Exporter, reloadCh chan<- actionMsg, path string) func(http.ResponseWriter, *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var result []byte
+
+		exporter.Logger().Info(
+			"received /loglevel", "client", r.RemoteAddr)
+
+		accept_type := r.Header.Get(acceptHeader)
+		if strings.Contains(accept_type, applicationJSON) {
+			accept_type = applicationJSON
+		} else if strings.Contains(accept_type, textPLAIN) {
+			accept_type = textPLAIN
+		} else {
+			accept_type = textHTML
+		}
+
+		switch r.Method {
+		case "GET":
+			switch accept_type {
+			case textPLAIN:
+				w.Header().Set(contentTypeHeader, textPLAIN)
+				result = []byte(
+					fmt.Sprintf("loglevel is currently set to %s", exporter.GetLogLevel()))
+			case applicationJSON:
+				w.Header().Set(contentTypeHeader, applicationJSON)
+				result = []byte(
+					fmt.Sprintf("{\"loglevel\"=\"%s\"}", exporter.GetLogLevel()))
+				w.Header().Set(contentLengthHeader, fmt.Sprint(len(result)))
+			default:
+				w.Header().Set(contentTypeHeader, textHTML)
+				healthTemplate.Execute(w, &tdata{
+					ExporterName: exporter.Config().Globals.ExporterName,
+					MetricsPath:  metricsPath,
+					DocsUrl:      docsUrl,
+					Message:      fmt.Sprintf("loglevel is currently set to %s", exporter.GetLogLevel()),
+				})
+				return
+			}
+			w.Write(result)
+		case "POST":
+			ctxval, ok := r.Context().Value(ctxKey{}).(*ctxValue)
+			if !ok {
+				err := errors.New("invalid context received")
+				HandleError(http.StatusInternalServerError, err, metricsPath, exporter, w, r)
+				return
+
+			}
+			msg := actionMsg{
+				actiontype: ACTION_LOGLEVEL,
+				logLevel:   strings.ToLower(ctxval.path),
+				retCh:      make(chan error),
+			}
+			reloadCh <- msg
+			if err := <-msg.retCh; err != nil {
+				switch accept_type {
+				case textPLAIN:
+					w.Header().Set(contentTypeHeader, textPLAIN)
+					result = []byte(
+						fmt.Sprintf("OK loglevel set to %s", err.Error()))
+				case applicationJSON:
+					w.Header().Set(contentTypeHeader, applicationJSON)
+					result = []byte(
+						fmt.Sprintf("{\"loglevel\"=\"%s\"}", err.Error()))
+					w.Header().Set(contentLengthHeader, fmt.Sprint(len(result)))
+				default:
+					w.Header().Set(contentTypeHeader, textHTML)
+					healthTemplate.Execute(w, &tdata{
+						ExporterName: exporter.Config().Globals.ExporterName,
+						MetricsPath:  metricsPath,
+						DocsUrl:      docsUrl,
+						Message:      fmt.Sprintf("OK loglevel set to %s", err),
+					})
+					return
+				}
+				w.Write(result)
+			} else {
+				HandleError(http.StatusInternalServerError, errors.New("KO something wrong with increase loglevel"), metricsPath, exporter, w, r)
+			}
+		default:
+			exporter.Logger().Info(
+				"received invalid method on /loglevel", "client", r.RemoteAddr)
+			HandleError(http.StatusMethodNotAllowed, errors.New("this endpoint requires a GET or POST request"), metricsPath, exporter, w, r)
 			return
 		}
-		configTemplate.Execute(w, &tdata{
-			ExporterName: exporter.Config().Globals.ExporterName,
-			MetricsPath:  metricsPath,
-			DocsUrl:      docsUrl,
-			Config:       string(config),
-		})
 	}
 }
 
@@ -199,12 +337,26 @@ func StatusHandlerFunc(metricsPath string, exporter Exporter) func(http.Response
 			ReloadTime: exporter.GetReloadTime(),
 		}
 
-		statusTemplate.Execute(w, &tdata{
-			ExporterName: exporter.Config().Globals.ExporterName,
-			MetricsPath:  metricsPath,
-			DocsUrl:      docsUrl,
-			Version:      vinfos,
-		})
+		accept_type := r.Header.Get(acceptHeader)
+		if strings.Contains(accept_type, applicationJSON) {
+			res, err := json.Marshal(vinfos)
+			if err != nil {
+				HandleError(http.StatusBadRequest, err, metricsPath, exporter, w, r)
+				return
+			}
+			w.Header().Set(contentTypeHeader, applicationJSON)
+			w.Header().Set(contentLengthHeader, fmt.Sprint(len(res)))
+			w.WriteHeader(http.StatusOK)
+			w.Write(res)
+		} else {
+			w.Header().Set(contentTypeHeader, textHTML)
+			statusTemplate.Execute(w, &tdata{
+				ExporterName: exporter.Config().Globals.ExporterName,
+				MetricsPath:  metricsPath,
+				DocsUrl:      docsUrl,
+				Version:      vinfos,
+			})
+		}
 	}
 }
 
@@ -215,20 +367,41 @@ func TargetsHandlerFunc(metricsPath string, exporter Exporter) func(http.Respons
 		var err error
 		c := exporter.Config()
 		// for _, t := range c.Targets {
-		targets_cfg, err = yaml.Marshal(c.Targets)
-		if err != nil {
-			// content = nil
-			HandleError(0, err, metricsPath, exporter, w, r)
-			return
+		accept_type := r.Header.Get(acceptHeader)
+		if strings.Contains(accept_type, applicationJSON) {
+			type targets struct {
+				Tgs []*TargetConfig `json:"targets"`
+			}
+			tgs := &targets{
+				Tgs: c.Targets,
+			}
+			targets_cfg, err = json.Marshal(tgs)
+			if err != nil {
+				// content = nil
+				HandleError(0, err, metricsPath, exporter, w, r)
+				return
+			}
+			w.Header().Set(contentTypeHeader, applicationJSON)
+			w.Header().Set(contentLengthHeader, fmt.Sprint(len(targets_cfg)))
+			w.WriteHeader(http.StatusOK)
+			w.Write(targets_cfg)
+		} else {
+			targets_cfg, err = yaml.Marshal(c.Targets)
+			if err != nil {
+				// content = nil
+				HandleError(0, err, metricsPath, exporter, w, r)
+				return
+			}
+			// targets_cfg = append(targets_cfg, content...)
+			// }
+			w.Header().Set(contentTypeHeader, textHTML)
+			targetsTemplate.Execute(w, &tdata{
+				ExporterName: exporter.Config().Globals.ExporterName,
+				MetricsPath:  metricsPath,
+				DocsUrl:      docsUrl,
+				Targets:      string(targets_cfg),
+			})
 		}
-		// targets_cfg = append(targets_cfg, content...)
-		// }
-		targetsTemplate.Execute(w, &tdata{
-			ExporterName: exporter.Config().Globals.ExporterName,
-			MetricsPath:  metricsPath,
-			DocsUrl:      docsUrl,
-			Targets:      string(targets_cfg),
-		})
 	}
 }