+++ main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"net/http"
 	"os"
@@ -15,12 +16,11 @@ import (
 	_ "net/http/pprof"
 
 	kingpin "github.com/alecthomas/kingpin/v2"
-	"github.com/go-kit/log/level"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"github.com/prometheus/common/expfmt"
-	"github.com/prometheus/common/promlog"
-	"github.com/prometheus/common/promlog/flag"
+	"github.com/prometheus/common/promslog"
+	"github.com/prometheus/common/promslog/flag"
 	"github.com/prometheus/common/version"
 	"github.com/prometheus/exporter-toolkit/web"
 	"github.com/prometheus/exporter-toolkit/web/kingpinflag"
@@ -36,7 +36,7 @@ var (
 	auth_key       = kingpin.Flag("auth_key", "In dry-run mode specify the auth_key to use, else ignored.").Short('a').String()
 	collector_name = kingpin.Flag("collector", "Specify the collector name restriction to collect, replace the collector_names set for each target.").Short('o').String()
 	toolkitFlags   = kingpinflag.AddFlags(kingpin.CommandLine, metricsPublishingPort)
-	logConfig      = promlog.Config{}
+	logConfig      = promslog.Config{}
 )
 
 const (
@@ -61,12 +61,32 @@ func newRoute(op int, path string, handler http.HandlerFunc) *route {
 	}
 
 }
-func BuildHandler(exporter Exporter) http.Handler {
+
+type ctxKey struct {
+}
+type ctxValue struct {
+	path string
+}
+
+type actionMsg struct {
+	actiontype int
+	logLevel   string
+	retCh      chan error
+}
+
+const (
+	ACTION_RELOAD   = iota
+	ACTION_LOGLEVEL = iota
+)
+
+func BuildHandler(exporter Exporter, actionCh chan<- actionMsg) http.Handler {
 	var routes = []*route{
-		newRoute(OpEgals, "/health", func(w http.ResponseWriter, r *http.Request) { http.Error(w, "OK", http.StatusOK) }),
 		newRoute(OpEgals, "/", HomeHandlerFunc(*metricsPath, exporter)),
-		newRoute(OpEgals, "/status", StatusHandlerFunc(*metricsPath, exporter)),
 		newRoute(OpEgals, "/config", ConfigHandlerFunc(*metricsPath, exporter)),
+		newRoute(OpEgals, "/health", func(w http.ResponseWriter, r *http.Request) { http.Error(w, "OK", http.StatusOK) }),
+		newRoute(OpMatch, "/loglevel(?:/(.*))?", LogLevelHandlerFunc(*metricsPath, exporter, actionCh, "")),
+		newRoute(OpEgals, "/reload", ReloadHandlerFunc(*metricsPath, exporter, actionCh)),
+		newRoute(OpEgals, "/status", StatusHandlerFunc(*metricsPath, exporter)),
 		newRoute(OpEgals, "/targets", TargetsHandlerFunc(*metricsPath, exporter)),
 		newRoute(OpEgals, *metricsPath, func(w http.ResponseWriter, r *http.Request) { ExporterHandlerFor(exporter).ServeHTTP(w, r) }),
 		// Expose exporter metrics separately, for debugging purposes.
@@ -82,8 +102,18 @@ func BuildHandler(exporter Exporter) http.Handler {
 				continue
 			}
 			if route.regex != nil {
-				if route.regex.MatchString(req.URL.Path) {
-					route.handler(w, req)
+				matches := route.regex.FindStringSubmatch(req.URL.Path)
+				if len(matches) > 0 {
+					var path string
+					if len(matches) > 1 {
+						path = matches[1]
+					}
+
+					ctxval := &ctxValue{
+						path: path,
+					}
+					ctx := context.WithValue(req.Context(), ctxKey{}, ctxval)
+					route.handler(w, req.WithContext(ctx))
 					return
 				}
 			} else if req.URL.Path == route.path {
@@ -96,6 +126,28 @@ func BuildHandler(exporter Exporter) http.Handler {
 	})
 }
 
+// ReloadHandlerFunc is the HTTP handler for the POST reload entry point (`/reload`).
+func ReloadHandlerFunc(metricsPath string, exporter Exporter, reloadCh chan<- actionMsg) func(http.ResponseWriter, *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
+		if r.Method != "POST" {
+			exporter.Logger().Info(
+				"received invalid method on /reload", "client", r.RemoteAddr)
+			HandleError(http.StatusMethodNotAllowed, errors.New("this endpoint requires a POST request"), metricsPath, exporter, w, r)
+			return
+		}
+		exporter.Logger().Info(
+			"received /reload from %s", "client", r.RemoteAddr)
+		msg := actionMsg{
+			actiontype: ACTION_RELOAD,
+			retCh:      make(chan error),
+		}
+		reloadCh <- msg
+		if err := <-msg.retCh; err != nil {
+			HandleError(http.StatusInternalServerError, err, metricsPath, exporter, w, r)
+		}
+		http.Error(w, "OK reload asked.", http.StatusOK)
+	}
+}
 func main() {
 
 	flag.AddFlags(kingpin.CommandLine, &logConfig)
@@ -108,9 +160,9 @@ func main() {
 		runtime.SetMutexProfileFraction(1)
 	}
 
-	logger := promlog.New(&logConfig)
-	level.Info(logger).Log("msg", fmt.Sprintf("Starting %s", exporter_name), "version", version.Info())
-	level.Info(logger).Log("msg", "Build context", "build_context", version.BuildContext())
+	logger := promslog.New(&logConfig)
+	logger.Info(fmt.Sprintf("Starting %s", exporter_name), "version", version.Info())
+	logger.Info("Build context", "build_context", version.BuildContext())
 
 	// Override the config.file default with the CONFIG environment variable, if set. If the flag is explicitly set, it
 	// will end up overriding either.
@@ -120,14 +172,14 @@ func main() {
 
 	exporter, err := NewExporter(*configFile, logger, *collector_name)
 	if err != nil {
-		level.Error(logger).Log("msg", fmt.Sprintf("Error creating exporter: %s", err))
+		logger.Error(fmt.Sprintf("Error creating exporter: %s", err.Error()))
 		os.Exit(1)
 	}
 
 	exporter.SetLogLevel(logConfig.Level.String())
 
 	if *dry_run {
-		level.Info(logger).Log("msg", "configuration OK.")
+		logger.Info("configuration OK.")
 		// get the target if defined
 		var (
 			err   error
@@ -144,18 +196,18 @@ func main() {
 					t_def, err := exporter.FindTarget(*model_name)
 					if err != nil {
 						err := fmt.Errorf("Target model '%s' not found: %s", *model_name, err)
-						level.Error(logger).Log(err.Error())
+						logger.Error(err.Error())
 						os.Exit(1)
 					}
 					if tmp_t, err = t_def.Config().Clone(*target_name, ""); err != nil {
 						err := fmt.Errorf("invalid url set for remote_target '%s' %s", *target_name, err)
-						level.Error(logger).Log(err.Error())
+						logger.Error(err.Error())
 						os.Exit(1)
 					}
 					t, err = exporter.AddTarget(tmp_t)
 					if err != nil {
 						err := fmt.Errorf("unable to create temporary target %s", err)
-						level.Error(logger).Log(err.Error())
+						logger.Error(err.Error())
 						os.Exit(1)
 					}
 					exporter.Config().Targets = append(exporter.Config().Targets, tmp_t)
@@ -169,14 +221,14 @@ func main() {
 			t, err = exporter.GetFirstTarget()
 		}
 		if err != nil {
-			level.Error(logger).Log("errmsg", err)
+			logger.Error(err.Error())
 			os.Exit(1)
 		}
 		if *auth_key != "" {
 			t.SetSymbol("auth_key", *auth_key)
 		}
 
-		level.Info(logger).Log("msg", fmt.Sprintf("try to collect target %s.", t.Name()))
+		logger.Info(fmt.Sprintf("try to collect target %s.", t.Name()))
 		timeout := time.Duration(0)
 		configTimeout := time.Duration(exporter.Config().Globals.ScrapeTimeout)
 
@@ -197,13 +249,13 @@ func main() {
 		gatherer := prometheus.Gatherers{exporter.WithContext(ctx, t)}
 		mfs, err := gatherer.Gather()
 		if err != nil {
-			level.Error(logger).Log("errmsg", fmt.Sprintf("Error gathering metrics: %v", err))
+			logger.Error(fmt.Sprintf("Error gathering metrics: %v", err))
 			if len(mfs) == 0 {
 				os.Exit(1)
 			}
 		}
 
-		level.Info(logger).Log("msg", "collect is OK. Dumping result to stdout.")
+		logger.Info("collect is OK. Dumping result to stdout.")
 
 		//dump metric to stdout
 		enc := expfmt.NewEncoder(os.Stdout, `text/plain; version=`+expfmt.TextVersion+`; charset=utf-8`)
@@ -211,7 +263,7 @@ func main() {
 		for _, mf := range mfs {
 			err := enc.Encode(mf)
 			if err != nil {
-				level.Error(logger).Log("Errmsg", err)
+				logger.Error(err.Error())
 				break
 			}
 		}
@@ -219,7 +271,7 @@ func main() {
 			// This in particular takes care of the final "# EOF\n" line for OpenMetrics.
 			closer.Close()
 		}
-		level.Info(logger).Log("msg", "dry-run is over. Exiting.")
+		logger.Info("dry-run is over. Exiting.")
 		os.Exit(0)
 	}
 
@@ -230,17 +282,38 @@ func main() {
 	signal.Notify(user2, syscall.SIGUSR2)
 	hup := make(chan os.Signal, 1)
 	signal.Notify(hup, syscall.SIGHUP)
+	actionCh := make(chan actionMsg)
 	go func() {
 		for {
 			select {
 			case <-user2:
-				exporter.IncreaseLogLevel()
+				exporter.IncreaseLogLevel("")
 			case <-hup:
-				level.Info(logger).Log("msg", "file reloading.")
+				logger.Info("file reloading.")
 				if err := exporter.ReloadConfig(); err != nil {
-					level.Error(logger).Log("msg", fmt.Sprintf("reload err: %s.", err))
+					logger.Info(fmt.Sprintf("reload err: %s.", err))
 				} else {
-					level.Info(logger).Log("msg", "file reloaded.")
+					logger.Info("file reloaded.")
+				}
+			case action := <-actionCh:
+				switch action.actiontype {
+				case ACTION_RELOAD:
+					logger.Info("file reloading received.")
+					if err := exporter.ReloadConfig(); err != nil {
+						logger.Error(fmt.Sprintf("reload err: %s.", err))
+						action.retCh <- err
+					} else {
+						logger.Info("file reloaded.")
+						action.retCh <- nil
+					}
+				case ACTION_LOGLEVEL:
+					if action.logLevel == "" {
+						logger.Info("increase loglevel received.")
+					} else {
+						logger.Info("set loglevel received.")
+					}
+					exporter.IncreaseLogLevel(action.logLevel)
+					action.retCh <- errors.New(exporter.GetLogLevel())
 				}
 			}
 		}
@@ -253,10 +326,10 @@ func main() {
 	go func() {
 		// Setup and start webserver.
 		server := &http.Server{
-			Handler: BuildHandler(exporter),
+			Handler: BuildHandler(exporter, actionCh),
 		}
 		if err := web.ListenAndServe(server, toolkitFlags, logger); err != nil {
-			level.Error(logger).Log("err", err)
+			logger.Error(err.Error())
 			os.Exit(1)
 		}
 	}()
@@ -264,7 +337,7 @@ func main() {
 	for {
 		select {
 		case <-term:
-			level.Info(logger).Log("msg", "Received SIGTERM, exiting gracefully...")
+			logger.Info("Received SIGTERM, exiting gracefully...")
 			os.Exit(0)
 		case <-srvc:
 			os.Exit(1)